Rename metadata operator to GetDeviceMetadata
The current name `GetMetadata` is much too vague and disrupts toolbox
search patterns, i.e. when typing the keyword "device" nothing will be
shown.

We do not include in the generator a provision for obsoleting the
previous operator, as this is a decision to make on a package-by-package
basis and we do not want to unnecessarily increase the surface area of
new interface packages.